rag/requirements.txt
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
